move disc and dirint from clearsky to irradiance
